refactor: split extensions
